+++ spec/hatchet/nltk_spec.rb
@@ -11,13 +11,13 @@ RSpec.describe 'NLTK corpora support' do
         expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX))
           remote: -----> Downloading NLTK corpora...
           remote: -----> Downloading NLTK packages: city_database stopwords
-          remote: .*: RuntimeWarning: 'nltk.downloader' found in sys.modules after import of package 'nltk', but prior to execution of 'nltk.downloader'; this may result in unpredictable behaviour
-          remote: \\[nltk_data\\] Downloading package city_database to
-          remote: \\[nltk_data\\]     /tmp/build_.*/.heroku/python/nltk_data...
-          remote: \\[nltk_data\\]   Unzipping corpora/city_database.zip.
-          remote: \\[nltk_data\\] Downloading package stopwords to
-          remote: \\[nltk_data\\]     /tmp/build_.*/.heroku/python/nltk_data...
-          remote: \\[nltk_data\\]   Unzipping corpora/stopwords.zip.
+          remote:        .*: RuntimeWarning: 'nltk.downloader' found in sys.modules after import of package 'nltk', but prior to execution of 'nltk.downloader'; this may result in unpredictable behaviour
+          remote:        \\[nltk_data\\] Downloading package city_database to
+          remote:        \\[nltk_data\\]     /app/.heroku/python/nltk_data...
+          remote:        \\[nltk_data\\]   Unzipping corpora/city_database.zip.
+          remote:        \\[nltk_data\\] Downloading package stopwords to
+          remote:        \\[nltk_data\\]     /app/.heroku/python/nltk_data...
+          remote:        \\[nltk_data\\]   Unzipping corpora/stopwords.zip.
         REGEX
 
         # TODO: Add a test that the downloaded corpora can be found at runtime.
@@ -44,7 +44,37 @@ RSpec.describe 'NLTK corpora support' do
 
     it 'does not try to install the specified NLTK corpora' do
       app.deploy do |app|
-        expect(app.output.downcase).not_to include('nltk')
+        expect(app.output).not_to include('NLTK')
+        expect(app.output).not_to include('nltk_data')
+      end
+    end
+  end
+
+  context 'when nltk.txt contains invalid entries' do
+    let(:app) { Hatchet::Runner.new('spec/fixtures/nltk_txt_invalid', allow_failure: true) }
+
+    it 'fails the build' do
+      app.deploy do |app|
+        expect(clean_output(app.output)).to match(Regexp.new(<<~REGEX, Regexp::MULTILINE))
+          remote: -----> Downloading NLTK corpora...
+          remote: -----> Downloading NLTK packages: invalid!
+          remote:        .+: RuntimeWarning: 'nltk.downloader' found in sys.modules after import of package 'nltk', but prior to execution of 'nltk.downloader'; this may result in unpredictable behaviour
+          remote:        \\[nltk_data\\] Error loading invalid!: Package 'invalid!' not found in
+          remote:        \\[nltk_data\\]     index
+          remote:        Error installing package. Retry\\? \\[n/y/e\\]
+          remote:        Traceback \\(most recent call last\\):
+          remote:        .+
+          remote:        EOFError: EOF when reading a line
+          remote: 
+          remote:  !     Error: Unable to download NLTK data.
+          remote:  !     
+          remote:  !     The 'python -m nltk.downloader' command to download NLTK
+          remote:  !     data did not exit successfully.
+          remote:  !     
+          remote:  !     See the log output above for more information.
+          remote: 
+          remote:  !     Push rejected, failed to compile Python app.
+        REGEX
       end
     end
   end